Rewrite queueing code
This is a big commit.

Instead of keeping images in an actual queue, just manage them with pointers
to places in the list.

There are 4 positions that matter:
1. The current image being viewed. This obviously moves as the user changes
   images. It serves as a "base" for the others.
2. The decode "head". This is incremented by the readahead thread after
   an image is ready for decode.
3. The decode "tail". This is incremented by the decode threads after they
   decode an image. The depth of the queue is the distance between the
   decode head and tail.
4. The "history tail". This lags behind the current image. It is incremented
   by the reclaim thread as it cleans up old images. If the decode head
   reaches the history tail, it indicates that the entire image set is
   in memory.

This seems to work pretty well. It keeps all the decode threads busy and
does a good job of making I/O the bottleneck.