Reallocate decoder thread surfaces
The decoder threads keep a surface around so that they do not have to
allocate and free them all the time. But bad things happen if the
surface is mismatched with the image being decoded.

Notice when the surface and image are mismatched. Free and reallocate
the surface to match when necessary.