Rework main loop
This is doing two things:

First, it switches from polling to sleeping for keyboard events. This
needs a little hackery because of the need to drain the event queue.
It will sleep for *one* event, then it polls to empty the rest of
the queue. Basically, it sleeps at _least_ once.

That uncovered another bug. If the first image doesn't render
correctly, no image is displayed. Fix that by having the render
function return success or failure.