Rewrite state and queueing code
This is a long time coming. It integrates the image->state states
with the lists. Images have one and only one state and are on one
and only one list.

Most transitions just move to a single state. Some transitions
are atomic, like the move from QUEUED_FOR_DECODE=>DECODING.

This definitely still needs some work. BRAND_NEW is probably broken,
for instance.