Re-render on mouse clicks
When the viewer is not in the foreground, it can be clicked to bring it
to the foreground. However, when this happends, the window needs to
redraw the area which might be revealed as it pops forward.

However, this is not currently happening. A "ghost" of the window that
*was* on top of it continues to be drawn.

The only events I see when this happens are mouse events. there do not
appear to be any focus or raise events to use.

Force a re-render when the window is clicked.